pyproject.toml
[tool.poetry]
name = "multi-tenant-saas"
version = "0.1.0"
description = "Production-grade multi-tenant SaaS backend — Stripe-style architecture"
authors = ["Projjwal"]
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.0"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
psycopg2-binary = "^2.9"
django-environ = "^0.11"
redis = "^5.0"
celery = "^5.3"
django-celery-beat = "^2.6"
argon2-cffi = "^23.1"
drf-spectacular = "^0.27"
structlog = "^24.1"
django-cors-headers = "^4.3"
factory-boy = "^3.3"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
isort = "^5.13"
flake8 = "^7.0"
mypy = "^1.10"
django-stubs = "^5.0"
djangorestframework-stubs = "^3.15"
pytest = "^8.0"
pytest-django = "^4.8"
pytest-cov = "^5.0"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
known_django = ["django"]
known_first_party = ["apps", "config"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
ignore_missing_imports = true
strict = false
warn_return_any = false

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
python_files = ["tests/test_*.py", "tests/**/test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
